+++ spec/support/shared_examples/a_compressible_trie.rb
@@ -15,12 +15,22 @@ shared_examples_for 'a compressible trie' do
   end
 
   context 'and the trie is compressed' do
-    before { trie.compress! }
+    let!(:original_root) { trie.root }
+    let!(:original_tree) { original_root.children_tree.keys }
+
+    before do
+      trie.compress!
+    end
 
     it_behaves_like 'a trie data structure'
 
     it 'is marked as compressed' do
       expect(trie).to be_compressed
+    end
+
+    it 'leaves the original root intact' do
+      expect(original_root.children_tree.keys).to eq original_tree
+      expect(trie.children_tree.keys).not_to eq original_tree
     end
   end
 end